+++ instrumentation/opentelemetry-instrumentation-django/src/opentelemetry/instrumentation/django/middleware/otel_middleware.py
@@ -315,6 +315,12 @@ class _DjangoMiddleware(MiddlewareMixin):
                     route = getattr(match, "route", None)
                     if route:
                         span.set_attribute(SpanAttributes.HTTP_ROUTE, route)
+                        duration_attrs = request.META[
+                            self._environ_duration_attr_key
+                        ]
+                        # Metrics currently use the 1.11.0 schema, which puts the route in `http.target`.
+                        # TODO: use `http.route` when the user sets `OTEL_SEMCONV_STABILITY_OPT_IN`.
+                        duration_attrs[SpanAttributes.HTTP_TARGET] = route
 
     def process_exception(self, request, exception):
         if self._excluded_urls.url_disabled(request.build_absolute_uri("?")):